password-hash: move traits to module
Clean up lib.rs by factoring trait definitions under a module
